Extract base config to dedicated module
As it interact with filesystem, it is moved to hardware subtree.
